pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel", "setuptools_scm[toml]>=6.2"]
build-backend = "setuptools.build_meta"

[project]
name = "pawnai"
dynamic = ["version"]
description = "Speaker diarization and transcription CLI tool"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [
    {name = "PawnAI Contributors"},
]
keywords = ["audio", "diarization", "transcription", "speaker", "embedding", "cli"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Multimedia :: Sound/Audio :: Speech",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

dependencies = [
    "typer>=0.9.0",
    "rich>=13.0.0",
    "nemo_toolkit[asr]>=2.6.0",
    "pyannote.audio==4.0.3",
    "lancedb>=0.1.0",
    "duckdb>=0.8.0",
    "pyarrow>=10.0.0",
    "prompt-toolkit>=3.0.0",
    "soundfile>=0.12.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
    "black>=23.0",
    "isort>=5.0",
    "flake8>=6.0",
    "mypy>=1.0",
]
docs = [
    "sphinx>=4.0",
    "sphinx-rtd-theme>=1.0",
]

[project.scripts]
pawnai = "pawnai.__main__:main"

[tool.setuptools.dynamic]
version = {attr = "pawnai.__version__"}

[project.urls]
Homepage = "https://github.com/example/pawnai"
Documentation = "https://github.com/example/pawnai#readme"
Repository = "https://github.com/example/pawnai.git"
"Bug Tracker" = "https://github.com/example/pawnai/issues"

[tool.setuptools]
packages = ["pawnai", "pawnai.core", "pawnai.cli", "pawnai.utils"]

[tool.black]
line-length = 100
target-version = ['py38']

[tool.isort]
profile = "black"
line_length = 100

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
addopts = "--cov=pawnai --cov-report=term-missing"
